--- src/css/team/team-individual-view.css
@@ -1,47 +1,0 @@
-.pro-image {
-    display: flex;
-    min-height: 15rem;
-}
-
-.image {
-    width: 250px;
-}
-
-.role-designation {
-    color: rgba(0, 0, 0, 0.4);
-    margin-top: -2%;
-    padding-bottom: 1em;
-}
-
-.short-biography {
-    margin-top: 1%;
-    /* padding-top: 1.2em; */
-    padding-bottom: 1.2em;
-}
-
-.social-links {
-    margin-top: 2.8%;
-}
-
-.f-link {
-    transition: all 0.3s;
-    font-size: 17px;
-    display: inline-block;
-    border-radius: 2px;
-    margin: 0.2em 0.6em 0 0;
-    padding: 0.3em;
-}
-
-.f-link:hover {
-    color: #f8d16c;
-    background-color: #293b52;
-}
-.divider {
-    padding-bottom: 1em;
-}
-.header {
-    padding-bottom: 1em !important;
-}
-.grid {
-    padding-bottom: 1em !important;
-}
