+++ src/css/team/tech-skills-card.css
@@ -1,5 +1,7 @@
 .label-container {
-    margin-top: 1%;
+    margin-top: 2%;
+    font-weight: 400;
+    font-size: 0.9rem;
 }
 
 .label {
@@ -9,7 +11,7 @@
     margin: 0 1.5% 1.3% 0;
     padding: 0.5em;
     border: none;
-    border-radius: 5%;
+    border-radius: 4rem;
 }
 
 .label-image {
